+++ l10n_br_sale_product/models/__init__.py
@@ -4,3 +4,4 @@
 
 from . import res_company
 from . import sale
+from . import sale_line
